Generate questionnaire docs

Generates a Quarto Markdown representation of the input for each specified language, renders it to Microsoft Word (DOCX) format and optionally uploads it as a Google Doc to the specified g_drive_dir.

Usage

gen_qstnr_docs(
  qstnr,
  survey_config,
  output_dir,
  langs = NULL,
  add_item_ids = TRUE,
  keep_qmd = TRUE,
  keep_docx = FALSE,
  create_g_docs = TRUE,
  g_drive_dir = NULL,
  quiet = FALSE
)

Arguments

qstnr

Unnested questionnaire tibble as returned by unnest_qstnr().

survey_config

Survey configuration list as returned by read_survey_config().

output_dir

Directory to write the generated questionnaire files to. A character scalar.

langs

Languages in which to output the questionnaire. A subset of survey_config$langs or NULL for all languages present.

add_item_ids

Whether or not to add questionnaire item identifiers next to the question texts.

keep_qmd

Whether to keep or delete the intermediate generated Quarto Markdown files.

keep_docx

Whether to keep or delete the intermediate generated Microsoft Word files.

create_g_docs

Whether or not to turn the generated files into Google Docs placed under the Google Drive path g_drive_dir.

g_drive_dir

Google Drive path under which to place the generated Google Docs. By default, the file is created in the current user's "My Drive" root folder. Ignored if create_g_docs = FALSE.

quiet

Whether or not to suppress printing status output from internal processing.

Value

output_dir, invisibly.

See also