Get survey group title

Extracts the multilanguage group title list for the specified group from the given survey config.

Usage

group_title(survey_config, group_id)

Arguments

survey_config

Survey configuration list as returned by read_survey_config().

group_id

Survey group identifier to retrieve the group title of. A character scalar.

Value

A list.

See also