qstnr: Helps Creating and Maintaining Survey Questionnaires

qstnr provides a framework to create and maintain survey questionnaires written in the TOML file format.

Package configuration

Some of qstnr's functionality is controlled via package-specific global configuration which can either be set via R options or environment variables (the former take precedence). This configuration includes:

Description	R option	Environment variable	Default value

Maximal timespan to preserve the package's pkgpins cache. Cache entries older than this will be deleted upon package loading.	qstnr.global_max_cache_age	QSTNR_GLOBAL_MAX_CACHE_AGE	"1 year"

Google Cloud Platform (GCP) Service Account Key file (in JSON format) to authorize the googledrive package to access and manage files on your Google Drive. See the relevant googledrive documentation for details.	qstnr.gcp_key	QSTNR_GCP_KEY

See also

Author

Maintainer: Salim Brüggemann salim-b@pm.me (ORCID)