Read in survey config

Assembles a raw TOML survey configuration into a combined survey configuration list.

Usage

read_survey_config(path, validate = NULL)

Arguments

path

Path to the survey config TOML file. A character scalar.

validate

Whether or not to validate all input TOML files, i.e. ensure that they adhere to their respective schemas. Validation usually adds a noticeable delay. If NULL, validation is performed only for files whose content has changed since the last successful validation.

Value

A list containing the combined survey configuration.

See also